# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountDeleteResponse": {
        "properties": {
          "deleted_at": {
            "title": "Deleted At",
            "type": "string"
          },
          "ok": {
            "const": true,
            "title": "Ok"
          },
          "removed": {
            "title": "Removed",
            "type": "object"
          }
        },
        "required": [
          "ok",
          "deleted_at",
          "removed"
        ],
        "title": "AccountDeleteResponse",
        "type": "object"
      },
...
      "DeleteAccountBody": {
        "properties": {
          "confirmation": {
            "maxLength": 20,
            "minLength": 6,
            "title": "Confirmation",
            "type": "string"
          }
        },
        "required": [
          "confirmation"
        ],
        "title": "DeleteAccountBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LegacyPrivacyDeleteRetiredResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "error": {
            "const": "legacy_privacy_endpoint_retired",
            "title": "Error"
          }
        },
        "required": [
          "error",
          "detail"
        ],
        "title": "LegacyPrivacyDeleteRetiredResponse",
        "type": "object"
      },
...
      "PrivacyDeleteResponse": {
        "description": "Legacy development fixture; not the current account-erasure contract.\n\nProduction returns HTTP 410. Use ``POST /api/me/delete`` for the\nauthenticated account-bound deletion flow.",
        "properties": {
          "deleted_at": {
            "title": "Deleted At",
            "type": "string"
          },
          "email_confirmation": {
            "enum": [
              "sent",
              "skipped"
            ],
            "title": "Email Confirmation",
            "type": "string"
          },
          "ok": {
            "const": true,
            "title": "Ok"
          },
          "removed": {
            "$ref": "#/components/schemas/PrivacyRemovalCounts"
          }
        },
        "required": [
          "ok",
          "deleted_at",
          "removed",
          "email_confirmation"
        ],
        "title": "PrivacyDeleteResponse",
        "type": "object"
      },
...
      "PrivacyRemovalCounts": {
        "description": "Per-store removal counts from the retired development fixture.",
        "properties": {
          "connections_messages": {
            "minimum": 0.0,
            "title": "Connections Messages",
            "type": "integer"
          },
          "connections_prefs": {
            "minimum": 0.0,
            "title": "Connections Prefs",
            "type": "integer"
          },
          "error_log": {
            "minimum": 0.0,
            "title": "Error Log",
            "type": "integer"
          },
          "match_requests": {
            "minimum": 0.0,
            "title": "Match Requests",
            "type": "integer"
          },
          "mock_checkouts": {
            "minimum": 0.0,
            "title": "Mock Checkouts",
            "type": "integer"
          },
          "newsletter_subscribers": {
            "minimum": 0.0,
            "title": "Newsletter Subscribers",
            "type": "integer"
          },
          "referrals": {
            "minimum": 0.0,
            "title": "Referrals",
            "type": "integer"
          },
          "structural_fingerprints": {
            "minimum": 0.0,
            "title": "Structural Fingerprints",
            "type": "integer"
          }
        },
        "required": [
          "newsletter_subscribers",
          "mock_checkouts",
          "error_log",
          "structural_fingerprints",
          "match_requests",
          "referrals",
          "connections_messages",
          "connections_prefs"
        ],
        "title": "PrivacyRemovalCounts",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "Optional API key. Anonymous requests fall back to the free tier (60 req/min). Provide an X-API-Key header to access higher tiers.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Structural Team",
      "url": "https://structural.bytedance.city/"
    },
    "description": "跨领域结构同构搜索引擎 — Structural Isomorphism API.\n\nPublic surface for the Structural cross-domain knowledge engine. Endpoints are grouped by capability (Ask / Phase / Phenomenon / Pricing / Newsletter / Admin). Rate limits are tier-based; supply an `X-API-Key` header to promote a request beyond the anonymous free-tier bucket.",
    "license": {
      "name": "Proprietary"
    },
    "title": "Structural",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/me/delete": {
      "post": {
        "description": "Erase registry assets, invalidate credentials and clear the cookie.",
        "operationId": "delete_my_account_api_me_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Permanently delete authenticated account",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/privacy/delete": {
      "delete": {
        "deprecated": true,
        "description": "Development compatibility endpoint. A well-formed, constraint-valid production request returns HTTP 410 before business logic or persistence; a malformed, overlong, or otherwise constraint-invalid query returns HTTP 422 before the handler. Signed-in users delete account data through /api/me/delete.",
        "operationId": "delete_data_api_privacy_delete_delete",
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Email"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyPrivacyDeleteRetiredResponse"
                }
              }
            },
            "description": "Production legacy deletion is retired"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Legacy email-code deletion (retired in production)",
        "tags": [
          "privacy"
        ]
      }
    },
...
  },
...
}